Make luna deterministic
We now apply the "advance universe" consistently,instead of ad-hoc
whenever a frame comes in. This means even if the framerate changes
the physics behaviour should be consistent.

To achieve thism `advanceUniverse` is called as many times as
necessary each frame, resulting in acceleration (and thus resulting
forces and positions) that are independent of framerate.